price-monitor/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
pandas>=2.0.0
streamlit>=1.28.0
schedule>=1.2.0
python-dotenv>=1.0.0
lxml>=4.9.0
webdriver-manager>=4.0.0
